chore(ci): cover more optional feature gates
