+++ .github/workflows/ci.yml
@@ -81,6 +81,8 @@ jobs:
       matrix:
         features:
           - "--no-default-features"
+          - "--features memory-embedding"
+          - "--features screenshot"
           - "--features panel"
           - "--features tool-pdf"
           - "--features memory-hnsw"
@@ -88,6 +90,13 @@ jobs:
           - "--features google"
           - "--features provider-vertex"
           - "--features whatsapp-web"
+          - "--features hardware"
+          - "--features peripheral-rpi"
+          - "--features probe"
+          - "--features android"
+          - "--features sandbox-landlock"
+          - "--features sandbox-firejail"
+          - "--features sandbox-bubblewrap"
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
       - uses: dtolnay/rust-toolchain@stable
